requirements.txt
# Automatically generated by https://github.com/damnever/pigar.


pymodbus==2.5.3
pyserial==3.5
PySide6==6.8.2.1
